Refactor editor subdirectory as layout specific
