+++ editor/assets/stylesheets/main.scss
@@ -1,6 +1,9 @@
 @import './variables';
 @import '../../inserter/style';
-@import '../../editor/style';
+@import '../../layout/style';
+@import '../../layout/mode-switcher/style';
+@import '../../modes/text-editor/style';
+@import '../../modes/visual-editor/style';
 
 body.toplevel_page_gutenberg {
 	background: #fff;
