+++ editor/layout/mode-switcher/style.scss
@@ -1,18 +1,8 @@
-/**
- * General Layout
- */
-
-.editor__header {
-	height: 56px;
-	padding: 10px;
-	border-bottom: 1px solid $light-gray-500;
-}
-
-.editor-mode-switcher {
+.mode-switcher {
 	position: relative;
 }
 
-.editor-mode-switcher__toggle {
+.mode-switcher__toggle {
 	color: $dark-gray-500;
 	padding: 8px;
 	align-items: center;
@@ -25,7 +15,7 @@
 	outline: none;
 }
 
-.editor-mode-switcher__content {
+.mode-switcher__content {
 	position: absolute;
 	top: 40px;
 	border: 1px solid $light-gray-500;
@@ -43,7 +33,7 @@
 	}
 }
 
-.editor-mode-switcher__arrow {
+.mode-switcher__arrow {
 	border: 10px dashed $light-gray-500;
 	height: 0;
 	line-height: 0;
@@ -69,37 +59,5 @@
 		border-top: none;
 		border-left-color: transparent;
 		border-right-color: transparent;
-	}
-}
-
-.editor-mode-text {
-	margin: 60px auto 0;
-	width: 700px;
-
-	textarea {
-		width: 700px;
-		border: none;
-		outline: none;
-		box-shadow: none;
-		resize: none;
-		padding: 0;
-		overflow: hidden;
-		font-family: $editor-html-font;
-	}
-}
-
-.editor-mode-visual {
-	margin: 60px auto 0;
-	width: 700px;
-
-	.mce-content-body {
-		font-family: $editor-font;
-		font-size: $editor-font-size;
-		line-height: $editor-line-height;
-
-		p {
-			font-size: $editor-font-size;
-			line-height: $editor-line-height;
-		}
 	}
 }
