+++ editor/layout/style.scss
@@ -1,0 +1,5 @@
+.layout__header {
+	height: 56px;
+	padding: 10px;
+	border-bottom: 1px solid $light-gray-500;
+}
