+++ editor/modes/text-editor/style.scss
@@ -1,0 +1,12 @@
+.text-editor {
+	display: block;
+	margin: 60px auto 0;
+	width: 700px;
+	border: none;
+	outline: none;
+	box-shadow: none;
+	resize: none;
+	padding: 0;
+	overflow: hidden;
+	font-family: $editor-html-font;
+}
