+++ editor/modes/visual-editor/style.scss
@@ -1,0 +1,15 @@
+.visual-editor {
+	margin: 60px auto 0;
+	width: 700px;
+
+	.mce-content-body {
+		font-family: $editor-font;
+		font-size: $editor-font-size;
+		line-height: $editor-line-height;
+
+		p {
+			font-size: $editor-font-size;
+			line-height: $editor-line-height;
+		}
+	}
+}
